+++ packages/monika-notification/channel/index.ts
@@ -1,3 +1,27 @@
+/**********************************************************************************
+ * MIT License                                                                    *
+ *                                                                                *
+ * Copyright (c) 2021 Hyperjump Technology                                        *
+ *                                                                                *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy   *
+ * of this software and associated documentation files (the "Software"), to deal  *
+ * in the Software without restriction, including without limitation the rights   *
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell      *
+ * copies of the Software, and to permit persons to whom the Software is          *
+ * furnished to do so, subject to the following conditions:                       *
+ *                                                                                *
+ * The above copyright notice and this permission notice shall be included in all *
+ * copies or substantial portions of the Software.                                *
+ *                                                                                *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR     *
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,       *
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE    *
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER         *
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,  *
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE  *
+ * SOFTWARE.                                                                      *
+ **********************************************************************************/
+
 import type { AnySchema } from 'joi'
 
 import * as dingtalk from './dingtalk'
@@ -6,15 +30,15 @@ import * as desktop from './desktop'
 import * as googlechat from './googlechat'
 import * as gotify from './gotify'
 import * as lark from './lark'
-// import * as mailgun from './mailgun'
-// import * as monikaNotif from './monika-notif'
-// import * as opsgenie from './opsgenie'
+import * as mailgun from './mailgun'
+import * as monikaNotif from './monika-notif'
+import * as opsgenie from './opsgenie'
 // import * as pagerduty from './pagerduty'
-// import * as pushbullet from './pushbullet'
-// import * as pushover from './pushover'
-// import * as sendgrid from './sendgrid'
-// import * as slack from './slack'
-// import * as smtp from './smtp'
+import * as pushbullet from './pushbullet'
+import * as pushover from './pushover'
+import * as sendgrid from './sendgrid'
+import * as slack from './slack'
+import * as smtp from './smtp'
 // import * as teams from './teams'
 // import * as telegram from './telegram'
 // import * as webhook from './webhook'
@@ -82,15 +106,15 @@ export const channels: Record<string, NotificationChannel> = {
   'google-chat': googlechat,
   gotify,
   lark,
-  // mailgun,
-  // 'monika-notif': monikaNotif,
-  // opsgenie,
+  mailgun,
+  'monika-notif': monikaNotif,
+  opsgenie,
   // pagerduty,
-  // pushbullet,
-  // pushover,
-  // sendgrid,
-  // slack,
-  // smtp,
+  pushbullet,
+  pushover,
+  sendgrid,
+  slack,
+  smtp,
   // teams,
   // telegram,
   // webhook,
