+++ packages/monika-notification/channel/slack.ts
@@ -24,7 +24,7 @@
 
 import Joi from 'joi'
 import type { NotificationMessage } from './'
-import { sendHttpRequest } from '../../../utils/http'
+import { sendHttpRequest } from '../utils/http'
 
 type NotificationData = {
   url: string
