+++ packages/monika-notification/package.json
@@ -17,11 +17,13 @@
   },
   "homepage": "https://github.com/raosan/monika#readme",
   "dependencies": {
+    "@sendgrid/mail": "^7.4.2",
     "axios": "^1.3.4",
     "date-fns": "^2.23.0",
     "form-data": "^4.0.0",
     "joi": "^17.4.0",
     "mailgen": "^2.0.15",
     "mailgun.js": "^3.5.4",
+    "nodemailer": "^6.5.0",
     "pino": "6.11.2"
   }}
